--- src/main/java/com/example/sns_project/interfaces/comment/CommentEdit.java
@@ -1,4 +1,0 @@
-package com.example.sns_project.interfaces.comment;
-
-public record CommentEdit(String comment) {
-}
