+++ src/main/java/com/example/sns_project/interfaces/comment/dto/CommentEdit.java
@@ -1,0 +1,4 @@
+package com.example.sns_project.interfaces.comment.dto;
+
+public record CommentEdit(String comment) {
+}
